+++ frontend/src/app/components/elements/timeline-template/timeline-template.component.ts
@@ -173,4 +173,8 @@ export class TimelineTemplateComponent implements OnInit {
                 .updateTimelines(this.timelineTemplate().id!)
                 .subscribe((result: any) => {});
     }
+
+    togglePresentation() {
+        this.timelineTemplate().presentation = !this.timelineTemplate().presentation;
+    }
 }